requirements.txt
fastapi==0.89.1
pydantic==1.10.4
PyJWT==2.6.0
PyMySQL==1.0.2
SQLAlchemy==1.3.13
uvicorn==0.20.0
